requirements.txt
python-dotenv>=1.0.0
aiohttp>=3.8.0
pandas>=2.0.0
matplotlib>=3.7.0
cachetools>=5.3.0